requirements.txt
anthropic>=0.28.0
httpx>=0.27.0
redis>=5.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pydantic>=2.7.0
structlog>=24.1.0
prometheus-client>=0.20.0
tenacity>=8.2.0
uvicorn[standard]>=0.29.0
fastapi>=0.111.0
